requirements.txt
Django
djangorestframework
selenium
beautifulsoup4
requests
fake-useragent
lxml
sentry_sdk